requirements.txt
asgiref==3.2.7
Django==3.0.5
mysqlclient==1.4.6
Pillow==7.1.1
pytz==2019.3
sqlparse==0.3.1
google-cloud-storage